+++ src/test/java/domain/participant/player/MoneyTest.java
@@ -1,6 +1,5 @@
-package domain.participant;
+package domain.participant.player;
 
-import domain.participant.player.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
